% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingrowth-package.R
\docType{package}
\name{ingrowth-package}
\alias{ingrowth-package}
\alias{_PACKAGE}
\title{ingrowth: A package for ingrowth imputation}
\description{
This mini-package contains only one function, \code{\link{ingrowth}}.
There is test data in \code{\link{treeData}} and \code{\link{plotData}}
}
\references{
García, O. (2021)"Imputing Ingrowth in Even-aged Permanent
Sample Plots". \emph{Mathematical and Computational Forestry
& Natural-Resource Sciences (MCFNS), 13}(2): 1--4.
(\url{http://mcfns.net/index.php/Journal/article/view/13.5})
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ogarciav/ingrowth}
  \item Report bugs at \url{https://github.com/ogarciav/ingrowth/issues}
}

}
\author{
\strong{Maintainer}: Oscar García \email{garcia@dasometrics.net} (\href{https://orcid.org/0000-0002-8995-1341}{ORCID})

}
\keyword{internal}
