% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\docType{data}
\name{plotData}
\alias{plotData}
\title{Test plot data set}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{PlotId}{Plot identifier}
\item{Thresh}{Minimum diameter threshold}
}
}
\usage{
plotData
}
\description{
Plot-level information from 3 permanent sample plots.
}
\seealso{
\code{\link{treeData}}
}
\keyword{datasets}
