% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeData.R
\docType{data}
\name{treeData}
\alias{treeData}
\title{Test tree data set}
\format{
A data frame with 1743 rows and 8 variables:
\describe{
\item{TreeId}{Tree identifier}
\item{MeasId}{Measurement identifier}
\item{PlotId}{Plot identifier}
\item{DBH}{Breast-height diameter}
\item{Id, Ht, HtC, CrwnClss}{Other variables not directly used by
\code{\link{ingrowth}}}
}
}
\usage{
treeData
}
\description{
Tree records from 3 permanent sample plots.
}
\seealso{
\code{\link{plotData}}
}
\keyword{datasets}
